[refactor] all the classes earlier in odemis.gui.comp.stream.py have been split into individual python files.
The stream file has been split into stream_bar and stream_panel files. The fastem specific code unrelated to the stream has also been sperated.
